refactor: add bot, create logger, and use  environments variables
- Refactor `BlueskyBot` into an async class;
- Create `Logger` class to manage task flow;
- Add environment variables;
- Add exception handling to catch any errors;
- Reinstate the webservice for deployment.

The `post` method in the `BlueskyBot` class was refactored to an async function to work correctly in the `process_queue` task.

A `Logger` class was created to handle task flow.

New environment variables `QUEUE_MAX_SIZE` and `PROCESS_INTERVAL` were added for better system control, avoiding hardcoded values.

Verifications were added before system initialization, and it will stop if any issues with environment variables are detected.

The webservice was reinstated to allow for deployment on Render.com.